# Skip already lowercased runes on transformation.